refactor: co-locate file type sniffing helpers in utils
